pyproject.toml
[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=0.5",
    "docx2txt>=0.8",
    "langchain>=1.0",
    "langchain-classic>=1.0",
    "langchain-chroma>=0.1",
    "langchain-community>=0.3",
    "langchain-core>=0.3",
    "langchain-ollama>=0.2",
    "langchain-openai>=0.2",
    "langchain-text-splitters>=0.3",
    "msal>=1.28",
    "openpyxl>=3.1",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
    "requests>=2.32",
    "streamlit>=1.38",
    "urllib3>=2.2",
]

[tool.setuptools]
packages = ["rag_agent", "rag_agent.rag", "rag_agent.scripts", "rag_agent.sharepoint"]

[tool.setuptools.package-dir]
"rag_agent" = "."
